Pull up PushDownAndCombineLimits
